Updated job controller tests.
