+++ job-service/server/src/test/java/org/verapdf/webapp/jobservice/server/controller/JobTaskResultHandlerTests.java
@@ -31,7 +31,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @AutoConfigureMockMvc
 @ActiveProfiles("test")
-public class JobTaskResultHandlerTest {
+public class JobTaskResultHandlerTests {
 	@Autowired
 	private MockMvc mockMvc;
